% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPFPCA-package.R
\docType{package}
\name{PPFPCA-package}
\alias{PPFPCA-package}
\title{PPFPCA}
\description{
It builds an algorithm to identify the occurrence of event outcome from trajectories of several predictors.
}
\references{
Wu, S., Müller, H., & Zhang, Z. (2013). FUNCTIONAL DATA ANALYSIS FOR POINT PROCESSES WITH RARE EVENTS. Statistica Sinica, 23(1), 1-23.
}
\author{
Liang Liang, Tianxi Cai, Hajime Uno
Maintainer: Hajime Uno <huno at jimmy.harvard.edu>
}
